Adição das imagens na tela de cadastro

+++ Texugos-PI/src/main/java/com/projetoIntegrador4Texugos/projetoIntegrador4/model/Produto.java
@@ -26,7 +26,7 @@ public class Produto {
 	private String descricaoProduto;
 	private double precoUnitProd;
 	private int qtdeEstoqueProd;
-	private int avaliacao;
+	private double avaliacao;
 	
 	private String pathImagem;
